removed size_t warnings

+++ src/Controller.cpp
@@ -3,7 +3,7 @@
 void Controller::map() {
     mapping = true;
     while (mapping) {
-        for (int i=0; i < robots.size(); i++) {
+        for (size_t i=0; i < robots.size(); i++) {
             robots[i]->mapGraph(graphMap, visitedNodes);
         }
     }
@@ -11,7 +11,7 @@ void Controller::map() {
 }
 
 bool Controller::visited(GraphNode* node, std::vector<GraphNode*> visited) {
-    for (int i = 0; i < visited.size(); i++) {
+    for (size_t i = 0; i < visited.size(); i++) {
         if (visited[i]->NodeID == node->NodeID) {
             return true;
         }
